Revert "Rename types/kernel_api_types/range_types to types/kernel_api/ranges"
This reverts commit cfb846b8110d115d320863ebdfe3222050726414.
